Add support for addons
Before template colocation, addons had to explicitly import a template
and set it via the `layout` property or using the `layout` decorator.
With template colocation, those can all be dropped: the build pipeline
automatically sets it up correctly (though it can also still be set up
manually if users so desire, using `setComponentTemplate`, which is what
colocation transforms to in the end).

Accordingly, running the codemod against an addon does more than it does
in the (default) app case. As with apps, it *does* colocate the template
with the backing class. However, it also looks up the imports for the
template in question and removes those along with their binding to the
`layout` property of a component class or via the `layout` decorator
from `ember-decorators`.

Notes on the change:

- Templates shared by multiple components are excluded. In this case, it
  is up to end users how to list those; they are reported in the same
  way other skipped templates (like partials) are.

- App components using the `layout` property or decorator are now
  updated as well. This is very rare, but the codemod can handle it
  trivially, so it does.

Additional improvements made along the way:

- Add basic JSDoc type annotations for the `Migrator` class to make
  working with its properties easier (autocomplete etc.).

- Replace `find` methods with getters for repeated lookups where
  possible (some have to be rerun after other transforms).

- Reuse the values looked up via the utility types.